refactor: 서비스 인터페이스와 구현을 하나의 클래스로 통합
CompanyServiceImpl을 제거하고, 서비스 메서드를 CompanyService 클래스에 직접 통합
인터페이스 계층을 없애고 서비스 구조를 단순화
